refactor: - 피드백 반영 - 패키지 분리 기준을 step에서 미션으로 변경
