+++ src/main/java/edu/nextstep/camp/carracing/Car.java
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.racingcar.step3;
+package edu.nextstep.camp.carracing;
 
 public class Car {
     private static final int MOVE_THRESHOLD = 4;
@@ -21,10 +21,6 @@ public class Car {
     }
 
     public String getPositionString() {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < position; i++) {
-            sb.append(MOVE_SYMBOL);
-        }
-        return sb.toString();
+        return MOVE_SYMBOL.repeat(position);
     }
 }
