+++ src/test/java/edu/nextstep/camp/tutorial/SetTest.java
@@ -1,3 +1,5 @@
+package edu.nextstep.camp.tutorial;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
