+++ src/test/java/edu/nextstep/camp/tutorial/StringTest.java
@@ -1,3 +1,5 @@
+package edu.nextstep.camp.tutorial;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
